Add auth checks

+++ app/(chat)/api/document/route.ts
@@ -1,35 +1,47 @@
-import { auth } from "@/app/(auth)/auth";
+import { auth } from '@/app/(auth)/auth';
 import {
   deleteDocumentsByIdAfterTimestamp,
   getDocumentsById,
   saveDocument,
-} from "@/db/queries";
+} from '@/db/queries';
 
 export async function GET(request: Request) {
   const { searchParams } = new URL(request.url);
-  const id = searchParams.get("id");
+  const id = searchParams.get('id');
 
   if (!id) {
-    return new Response("Missing id", { status: 400 });
+    return new Response('Missing id', { status: 400 });
+  }
+
+  const session = await auth();
+
+  if (!session || !session.user) {
+    return new Response('Unauthorized', { status: 401 });
   }
 
   const documents = await getDocumentsById({ id });
+
+  const [document] = documents;
+
+  if (document.userId !== session.user.id) {
+    return new Response('Unauthorized', { status: 401 });
+  }
 
   return Response.json(documents, { status: 200 });
 }
 
 export async function POST(request: Request) {
   const { searchParams } = new URL(request.url);
-  const id = searchParams.get("id");
+  const id = searchParams.get('id');
 
   if (!id) {
-    return new Response("Missing id", { status: 400 });
+    return new Response('Missing id', { status: 400 });
   }
 
   const session = await auth();
 
   if (!session) {
-    return new Response("Unauthorized", { status: 401 });
+    return new Response('Unauthorized', { status: 401 });
   }
 
   const { content, title }: { content: string; title: string } =
@@ -45,18 +57,32 @@ export async function POST(request: Request) {
 
     return Response.json(document, { status: 200 });
   } else {
-    return new Response("Unauthorized", { status: 401 });
+    return new Response('Unauthorized', { status: 401 });
   }
 }
 
 export async function PATCH(request: Request) {
   const { searchParams } = new URL(request.url);
-  const id = searchParams.get("id");
+  const id = searchParams.get('id');
 
   const { timestamp }: { timestamp: string } = await request.json();
 
   if (!id) {
-    return new Response("Missing id", { status: 400 });
+    return new Response('Missing id', { status: 400 });
+  }
+
+  const session = await auth();
+
+  if (!session || !session.user) {
+    return new Response('Unauthorized', { status: 401 });
+  }
+
+  const documents = await getDocumentsById({ id });
+
+  const [document] = documents;
+
+  if (document.userId !== session.user.id) {
+    return new Response('Unauthorized', { status: 401 });
   }
 
   await deleteDocumentsByIdAfterTimestamp({
@@ -64,5 +90,5 @@ export async function PATCH(request: Request) {
     timestamp: new Date(timestamp),
   });
 
-  return new Response("Deleted", { status: 200 });
+  return new Response('Deleted', { status: 200 });
 }
